+++ packages/date-picker/src/picker-content/single.tsx
@@ -1,321 +1,14 @@
-import {
-  forwardRef,
-  useCallback,
-  cloneElement,
-  useState,
-  useEffect,
-  useMemo,
-} from "react"
-import { throttleByRaf } from "@illa-design/system"
-import { Button } from "@illa-design/button"
-import dayjs, { Dayjs } from "dayjs"
-import { TimePickerPopup } from "@illa-design/time-picker"
-import { Calendar } from "@illa-design/calendar"
+import { forwardRef, } from "react"
 import {
   DatePickerProps,
   MonthPickerProps,
   YearPickerProps,
-  CommonSingleProps,
 } from "../interface"
 import { Picker } from "../picker"
-import {
-  triContentCommonCss,
-  applyShortContainerCss,
-  shortCutsCss,
-  singlePickerContentCss,
-  showTimeContainerCss,
-  showTimeHeaderCss,
-  popupCss,
-  horShortcuts,
-  vertShortcuts,
-  nowButtonCss,
-} from "../style"
-import { initFormat } from "../utils"
-
-const CommonPicker = forwardRef<HTMLDivElement, CommonSingleProps>(
-  (props, ref) => {
-    const {
-      type,
-      disabled,
-      allowClear = true,
-      position = "bottom-start",
-      placeholder = "",
-      shortcuts,
-      shortcutsPlacementLeft,
-      error,
-      size = "medium",
-      popupVisible,
-      onVisibleChange,
-      onChange,
-      onSelect,
-      onClear,
-      editable = true,
-      onSelectShortcut,
-      disabledDate,
-      defaultPickerValue,
-      showTime,
-      onOk,
-      format,
-      value,
-      defaultValue,
-      showNowBtn,
-      disabledTime,
-      colorScheme,
-      readOnly,
-      ...restProps
-    } = props
-
-    const tpProps = typeof showTime === "object" ? showTime : {}
-    const isBooleanShowTime = typeof showTime === "boolean" ? showTime : false
-
-    const finalFormat = format || initFormat(type, isBooleanShowTime)
-
-    const initValue =
-      value || defaultValue
-        ? dayjs(value || defaultValue).format(finalFormat as string)
-        : ""
-    const [inputVal, setInputVal] = useState<string>(initValue)
-    useEffect(() => {
-      if (value) {
-        const _initValue = dayjs(value).format(finalFormat as string)
-        setInputVal(_initValue)
-        setCalendarShortCuts(dayjs(value))
-      }
-    }, [value])
-    useEffect(() => {
-      // YYYY-MM-DD
-      if (finalFormat?.length) {
-        setInputVal(
-          inputVal ? dayjs(inputVal).format(finalFormat as string) : inputVal,
-        )
-        setCalendarShortCuts(inputVal ? dayjs(value) : undefined)
-      }
-    }, [finalFormat])
-
-    const [calendarShortCuts, setCalendarShortCuts] = useState<
-      Dayjs | "clear"
-    >()
-    const [showTrigger, setShowTrigger] = useState<boolean>(
-      popupVisible ?? false,
-    )
-    const mergedDefaultValue = value || defaultValue || defaultPickerValue
-    const showTimeMerged =
-      (isBooleanShowTime || Object.keys(tpProps).length > 0) && type === "day"
-
-    const [valueShow, setValueShow] = useState<Dayjs>()
-
-    const [calendarValue, setCalendarValue] = useState<Dayjs>(dayjs())
-
-    const finalValue = (calendar?: Dayjs | null, timePicker?: Dayjs | null) => {
-      calendar = calendar || calendarValue
-      timePicker = timePicker || dayjs()
-      return dayjs(
-        `${calendar.format("YYYY-MM-DD")} ${timePicker.format("HH:mm:ss")}`,
-      )
-    }
-
-    const showCalendarTodayButton = useMemo(() => {
-      // if show time select, hide Today button
-      if (showTimeMerged) {
-        return false
-      }
-      if (showNowBtn === undefined && type === "day") {
-        return true
-      } else {
-        return showNowBtn && !isBooleanShowTime && !shortcuts?.length
-      }
-    }, [])
-
-    const changeDate = (date: Dayjs, time?: Dayjs) => {
-      let value = finalValue(date, time)
-      let valueFormat = value.format(finalFormat as string)
-      onSelect?.(valueFormat, value)
-      if (!showTimeMerged) {
-        onChange?.(valueFormat, value)
-        setShowTrigger(false)
-        setInputVal(valueFormat)
-        setCalendarValue(value)
-        setValueShow(value)
-        setCalendarShortCuts(value)
-      } else {
-        setCalendarValue(date)
-        setValueShow(value)
-      }
-    }
-    const clearDate = () => {
-      setInputVal("")
-      setCalendarValue(dayjs())
-      setCalendarShortCuts("clear")
-      setValueShow(undefined)
-
-      onClear?.()
-      setShowTrigger(false)
-    }
-
-    const clickNow = () => {
-      let current = dayjs()
-      setInputVal(current.format(finalFormat as string))
-      setCalendarShortCuts(current)
-      onChange?.(current.format(finalFormat as string), current)
-      setShowTrigger(false)
-    }
-
-    const handleShortEnter = useCallback(
-      throttleByRaf((item: any) => {
-        if (item.value().isSame(calendarValue, "date")) return
-        setCalendarShortCuts(item.value())
-      }),
-      [calendarShortCuts],
-    )
-    const handleShortLeave = useCallback(
-      throttleByRaf((item: any) => {
-        if (item.value().isSame(calendarValue, "date")) return
-        setCalendarShortCuts("clear")
-      }),
-      [calendarShortCuts],
-    )
-
-    const onConfirmValue = (time: Dayjs) => {
-      let value = finalValue(calendarValue, time)
-      let valueFormat = value.format(finalFormat as string)
-      onChange?.(valueFormat, value)
-      onOk?.(valueFormat, value)
-      setShowTrigger(false)
-      setInputVal(valueFormat)
-      setCalendarShortCuts(value)
-      setCalendarValue(dayjs())
-      setValueShow(undefined)
-    }
-
-    function ShortcutsCompt() {
-      return shortcuts ? (
-        <div css={applyShortContainerCss(shortcutsPlacementLeft)}>
-          {/* eslint-disable-next-line react/prop-types */}
-          {shortcuts.map((item, key) => {
-            return (
-              <div
-                css={shortCutsCss}
-                key={key}
-                onMouseEnter={() => handleShortEnter(item)}
-                onMouseLeave={() => handleShortLeave(item)}
-                onClick={() => {
-                  setCalendarValue(item.value() as Dayjs)
-                  changeDate(item.value() as Dayjs)
-                  onSelectShortcut?.(item)
-                }}
-              >
-                {item.text}
-              </div>
-            )
-          })}
-        </div>
-      ) : null
-    }
-    console.log({
-      inputVal,
-      calendarShortCuts,
-      calendarValue: calendarValue?.format(finalFormat as string),
-      valueShowString: valueShow?.format(finalFormat as string),
-    })
-    return (
-      <Picker
-        ref={ref}
-        disabled={disabled}
-        allowClear={allowClear}
-        position={position}
-        placeholder={placeholder}
-        inputVal={valueShow?.format(finalFormat as string) || inputVal}
-        error={error}
-        size={size}
-        popupVisible={showTrigger}
-        editable={editable}
-        readOnly={readOnly}
-        colorScheme={colorScheme}
-        pickerContent={
-          <div css={singlePickerContentCss}>
-            {shortcutsPlacementLeft && (
-              <div css={vertShortcuts}>
-                <ShortcutsCompt />
-              </div>
-            )}
-            <div>
-              <Calendar
-                panel
-                isTodayTarget
-                mode={type}
-                panelTodayBtn={showCalendarTodayButton}
-                _css={triContentCommonCss}
-                onChange={(date: Dayjs) => {
-                  changeDate(date, valueShow)
-                }}
-                disabledDate={disabledDate}
-                defaultDate={calendarValue || dayjs(mergedDefaultValue)}
-                defaultSelectedDate={calendarShortCuts}
-              />
-              {(shortcuts || showTime) && (
-                <div css={horShortcuts}>
-                  {shortcuts && !shortcutsPlacementLeft ? (
-                    <ShortcutsCompt />
-                  ) : showNowBtn ? (
-                    <Button
-                      colorScheme="gray"
-                      css={nowButtonCss}
-                      onClick={() => {
-                        clickNow()
-                      }}
-                    >
-                      Now
-                    </Button>
-                  ) : null}
-                </div>
-              )}
-            </div>
-            {showTimeMerged && (
-              <div css={showTimeContainerCss}>
-                <div css={showTimeHeaderCss}>time</div>
-                <div css={popupCss}>
-                  {cloneElement(<TimePickerPopup />, {
-                    isRangePicker: false,
-                    format: "HH:mm:ss",
-                    valueShow: valueShow,
-                    setValueShow,
-                    popupVisible: showTrigger,
-                    onConfirmValue,
-                    showNowBtn: false,
-                    disabledHours: disabledTime?.().disabledHours,
-                    disabledMinutes: disabledTime?.().disabledMinutes,
-                    disabledSeconds: disabledTime?.().disabledSeconds,
-                    onSelect: (valueString: string, value: Dayjs) => {
-                      changeDate(calendarValue, value)
-                    },
-                    ...tpProps,
-                  })}
-                </div>
-              </div>
-            )}
-          </div>
-        }
-        onVisibleChange={(visible) => {
-          if (!visible) {
-            setValueShow(undefined)
-          }
-          onVisibleChange?.(visible)
-        }}
-        onChangeInputVal={setInputVal}
-        onChangeVisible={setShowTrigger}
-        onClearDate={clearDate}
-        onClear={onClear}
-        {...restProps}
-      />
-    )
-  },
-)
-
-CommonPicker.displayName = "CommonPicker"
 
 export const DatePicker = forwardRef<HTMLDivElement, DatePickerProps>(
   (props, ref) => {
-    return <CommonPicker {...props} ref={ref} type={"day"} />
+    return <Picker {...props} ref={ref} type={"day"} />
   },
 )
 
@@ -323,7 +16,7 @@ DatePicker.displayName = "DatePicker"
 
 export const MonthPicker = forwardRef<HTMLDivElement, MonthPickerProps>(
   (props, ref) => {
-    return <CommonPicker {...props} ref={ref} type={"month"} />
+    return <Picker {...props} ref={ref} type={"month"} />
   },
 )
 
@@ -331,7 +24,7 @@ MonthPicker.displayName = "MonthPicker"
 
 export const YearPicker = forwardRef<HTMLDivElement, YearPickerProps>(
   (props, ref) => {
-    return <CommonPicker {...props} ref={ref} type={"year"} />
+    return <Picker {...props} ref={ref} type={"year"} />
   },
 )
